+++ sdk/azidentity/CHANGELOG.md
@@ -6,7 +6,12 @@
 
 ### Breaking Changes
 > These changes affect only code written against a beta version such as v1.3.0-beta.4
+* Moved `NewWorkloadIdentityCredential()` parameters into `WorkloadIdentityCredentialOptions`.
+  The constructor now reads default configuration from environment variables set by the Azure
+  workload identity webhook by default.
+  ([#20478](https://github.com/Azure/azure-sdk-for-go/pull/20478))
 * Removed CAE support. It will return in the next beta release.
+  ([#20479](https://github.com/Azure/azure-sdk-for-go/pull/20479))
 
 ### Bugs Fixed
 * Fixed an issue in `DefaultAzureCredential` that could cause the managed identity endpoint check to fail in rare circumstances.